chore: move redis wrapper implementation to repository.redis package
